[Refactor] 희망 회사 도메인 패키지 구조를 정리한다 (#19)
